+++ plugins/02_TheBurningCrusade/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v125
+
+* Fix mount journal ids for Netherwing Drakes in Shadowmoon Valley
+
 # v99
 
 * Localization updates: esES