+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -1,6 +1,7 @@
 # v125
 
 * Add Translocated Gorger mount
+* Add additional Phase Orbs in K'aresh
 * Localization updates: esES, fr,FR, ruRU, zhCN
 
 # v124